requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
pyserial==3.5
